app/requirements.txt
sqlalchemy
psycopg2-binary
pandas
faker